added frontend and middleware

+++ graphql/resolvers/bookings.js
@@ -1,12 +1,15 @@
 const Booking = require('../../models/booking');
 const Event = require('../../models/event');
-const { transformBooking } = require('./merge')
+const { transformBooking, transformEvent } = require('./merge')
 
 module.exports = {
 
-    bookings: async () => {
+    bookings: async (args, req) => {
+        if(!req.isAuth){
+            throw new Error("Unauthenticated!")
+        }
         try{
-        const bookings = await Booking.find()
+        const bookings = await Booking.find({user: req.userId}) //
         return bookings.map(booking => {            
                 return transformBooking(booking)
             });
@@ -15,10 +18,13 @@ module.exports = {
         }
     },
     
-    bookEvent: async args => {
+    bookEvent: async (args, req) => {
+        if(!req.isAuth){
+            throw new Error("Unauthenticated!")
+        }
         const fetchedEvent = await Event.findOne({_id: args.eventId});
         const booking = new Booking({
-            user: '5eda9e7442d3ee6c0fac8aca',
+            user: req.userId,
             event: fetchedEvent
         });
         const result = await booking.save();
@@ -26,11 +32,19 @@ module.exports = {
     },
 
     //cancelBooking(bookingId: ID!): Event!
-    cancelBooking: async args => {
+    cancelBooking: async (args, req)=> {
+        if(!req.isAuth){
+            throw new Error("Unauthenticated!")
+        }
         try{
             const fetchedBooking = await Booking.findById({_id: args.bookingId}).populate('event');
             if(!fetchedBooking){
                 throw new Error("Booking does not exist")
+            }
+            console.log("fetchedBooking.user" + fetchedBooking.user)
+            console.log("req.userId" + req.userId)
+            if(fetchedBooking.user.toString() !== req.userId.toString()){
+                throw new Error("Can cancel only user's own bookings")
             }
             const event = transformEvent(fetchedBooking.event)
             await Booking.deleteOne({_id:args.bookingId})
